renamed cluster file back to resource-monitoring-server.{h,cpp}
