+++ examples/chef/common/clusters/resource-monitoring/chef-resource-monitoring-delegates.cpp
@@ -18,7 +18,7 @@
 #include <app-common/zap-generated/ids/Attributes.h>
 #include <app-common/zap-generated/ids/Clusters.h>
 #include <app/clusters/resource-monitoring-server/resource-monitoring-cluster-objects.h>
-#include <app/clusters/resource-monitoring-server/resource-monitoring-cluster.h>
+#include <app/clusters/resource-monitoring-server/resource-monitoring-server.h>
 #include <lib/core/TLVReader.h>
 #include <lib/support/BufferReader.h>
 #include <lib/support/BufferWriter.h>
