+++ examples/chef/common/clusters/resource-monitoring/chef-resource-monitoring-delegates.h
@@ -19,7 +19,7 @@
 #include <app-common/zap-generated/ids/Clusters.h>
 #include <app/clusters/resource-monitoring-server/CodegenIntegration.h>
 #include <app/clusters/resource-monitoring-server/resource-monitoring-cluster-objects.h>
-#include <app/clusters/resource-monitoring-server/resource-monitoring-cluster.h>
+#include <app/clusters/resource-monitoring-server/resource-monitoring-server.h>
 
 namespace chip {
 namespace app {
